Extract some of MethodLevelOptsTest into BoxUnboxTests for mo' betta review diff
